requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
tqdm>=4.66.0
pyyaml>=6.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
